Compile metrics into the lib.
The metrics of the built-in fonts are now actually built into
the lib.  Currently only widths of simple code-points in the
default encoding (which is not yet properly done when printing
strings).  Named characters for use with other encodings will
have to come later.

Also, bump version and add myself to authors.