Encode text as WinAnsiEncoding.
I don't do windows, but I do find the PDF /WinAnsiEncoding more
usefull than the default encoding.  Partly because it is so similar to
the first 255 code points of Unicode and partly because it features
åäö, which are used extensively in my native language.

Unfortunatley, the codepoints used in the .afm files are in the
default encoding and not in /WinAnsiEncoding, so that will need some
work (but it should be fixable compile-time).