Provide special encoding for Symbol font.
BuiltinFont::Symbol.get_encoding() now returns the SymbolEncoding that
PDF provides for that font rather than WinAnsiEncoding.  Another
encoding for the ZapfDingbats font remains to be implemented.

Prepare for other font sources by making FontSource a trait rather
than an enum, and making BuiltinFont an enum that provides the 14
builtin fonts and implements the FontSource trait.  Support for
TrueType and/or Type1 / Type3 fonts should be implemented as separate
structs implementing the FontSource trait.

WIN_ANSI_ENCODING is no longer exposed as a symbol, but only from
get_encoding() on the builtin fonts that uses it.